Made header & footer components more simple | Updates to component tests
