+++ libs/eo/landing-page/shell/src/lib/eo-landing-page.store.ts
