+++ libs/eo/shared/ui-page-templates/src/index.ts
@@ -14,8 +14,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './lib/ui-page-templates.module';
-
-// @todo: JK, understand -> Do we want to export these independently, or make them available ony via the module above?
-// export * from './lib/header/eo-header.component';
-// export * from './lib/header/eo-footer.component';
+export * from './lib/header/eo-header.component';
+export * from './lib/footer/eo-footer.component';
