+++ libs/eo/shared/ui-page-templates/src/lib/footer/eo-footer.component.spec.ts
@@ -14,41 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {
-  SpectacularAppComponent,
-  SpectacularFeatureTestingModule,
-} from '@ngworker/spectacular';
-import { render, screen } from '@testing-library/angular';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { render, screen } from '@testing-library/angular'
+import { EoFooterScam, EoFooterComponent } from './eo-footer.component';
 
-import { EoLandingPageShellModule } from './eo-landing-page-shell.module';
+describe(EoFooterComponent.name, () => {
 
-describe('EnergyOrigin landing page', () => {
-  const findEnergyOriginLogo = () =>
-    screen.findByRole('img', { name: 'EnergyOrigin' });
-
-  const findEnerginetLogo = () =>
-    screen.findByRole('img', { name: 'Energinet' });
+  const findEnerginetLogo = () => screen.findByRole('img', { name: 'Energinet' });
 
   const findPhoneLink = () => screen.findByRole('link', { name: /phone/i });
 
   const findEmailLink = () => screen.findByRole('link', { name: /email/i });
 
   beforeEach(async () => {
-    const { navigate } = await render(SpectacularAppComponent, {
-      imports: [
-        SpectacularFeatureTestingModule.withFeature({
-          featureModule: EoLandingPageShellModule,
-          featurePath: '',
-        }),
-        HttpClientTestingModule,
-      ],
+    await render(EoFooterComponent, {
+      imports: [EoFooterScam]
     });
-    await navigate('/');
-  });
-
-  it('displays the EnergyOrigin logo', async () => {
-    expect(await findEnergyOriginLogo()).toBeInTheDocument();
   });
 
   it('displays the Energinet logo', async () => {
@@ -62,4 +42,5 @@ describe('EnergyOrigin landing page', () => {
   it('displays an e-mail link', async () => {
     expect((await findEmailLink()).getAttribute('href')).toMatch(/^mailto:/);
   });
+
 });
