+++ libs/eo/shared/ui-page-templates/src/lib/footer/eo-footer.component.ts
@@ -108,4 +108,4 @@ export class EoFooterComponent {}
   declarations: [EoFooterComponent],
   exports: [EoFooterComponent],
 })
-export class EoFooterComponentScam {}
+export class EoFooterScam {}
