+++ libs/eo/shared/ui-page-templates/src/lib/header/eo-header.component.ts
@@ -18,15 +18,9 @@ import {
   ChangeDetectionStrategy,
   Component,
   NgModule,
-  ViewEncapsulation,
-  Input
+  ViewEncapsulation
 } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { MatButtonModule } from '@angular/material/button';
 import { MatToolbarModule } from '@angular/material/toolbar';
-import { LetModule } from '@rx-angular/template';
-import { EoHeaderStore } from './eo-header.store';
-import { Observable } from 'rxjs';
 
 const selector = 'eo-header';
 
@@ -36,8 +30,6 @@ const selector = 'eo-header';
   selector,
   styles: [
     `
-      @use '@energinet-datahub/watt/utils' as watt;
-
       .${selector}__toolbar {
         display: flex;
         justify-content: space-between;
@@ -49,33 +41,6 @@ const selector = 'eo-header';
 
         // Adjustment locally to contain a normal sized button
         height: var(--watt-space-xl);
-
-        // 1. Primary Watt Button.
-        // 2. Normal size Watt Button.
-        // 3. Custom size for Watt Button in App bar.
-        // 4. Align text vertically.
-        a {
-          @include watt.typography-watt-button; // [1]
-
-          --height: calc(10 * var(--watt-space-xs));
-          --inset-squish-m--x: var(--watt-space-m);
-          --inset-squish-m--y: var(--watt-space-s);
-
-          background: var(--watt-color-primary); // [1]
-          color: var(--watt-color-primary-contrast); // [1]
-
-          min-width: 6.25rem; // [2]
-          height: var(--height); // [3]
-          padding: var(--inset-squish-m--y) var(--inset-squish-m--x); // [3]
-
-          line-height: calc(
-            var(--height) - 2 * var(--inset-squish-m--y)
-          ); // [3] [4]
-
-          &:hover {
-            text-decoration: none; // [1]
-          }
-        }
       }
     `,
   ],
@@ -85,29 +50,15 @@ const selector = 'eo-header';
       class="${selector}__toolbar watt-space-inset-squished-m"
     >
       <img src="assets/energyorigin-logo.svg" alt="EnergyOrigin" />
-      <ng-container *ngIf="showStartButton">
-        <a
-          mat-button
-          mat-flat-button
-          *rxLet="loginUrl$ as loginUrl"
-          [href]="loginUrl"
-          >Start</a
-        >
-      </ng-container>
+      <ng-content></ng-content>
     </mat-toolbar>
   `,
 })
-export class EoHeaderComponent {
-  @Input() showStartButton = true;
-
-  loginUrl$: Observable<string> = this.headerStore.authenticationUrl$;
-  constructor(private headerStore: EoHeaderStore) {}
-}
+export class EoHeaderComponent { }
 
 @NgModule({
   declarations: [EoHeaderComponent],
   exports: [EoHeaderComponent],
-  imports: [CommonModule, MatToolbarModule, MatButtonModule, LetModule],
-  providers: [EoHeaderStore],
+  imports: [MatToolbarModule]
 })
-export class EoHeaderComponentScam {}
+export class EoHeaderScam {}
