+++ libs/eo/shared/ui-page-templates/src/test-setup.ts
@@ -16,7 +16,8 @@
  */
 import 'jest-preset-angular/setup-jest';
 import { addDomMatchers } from '@energinet-datahub/gf/test-util-matchers';
-import { setUpTestbed } from '@energinet-datahub/gf/test-util-staging';
+import { setUpTestbed, setUpAngularTestingLibrary } from '@energinet-datahub/gf/test-util-staging';
 
 addDomMatchers();
 setUpTestbed();
+setUpAngularTestingLibrary();
