+++ libs/ett/auth/data-access-api/src/lib/oidc/auth-http.service.ts
@@ -66,7 +66,9 @@ export class AuthHttp {
 
   // @todo: Remove <any> -> Return type for this request?
   // @todo: Expected payload for this request?
+  /*
   acceptTerms(): Observable<any> {
     return this.http.post(`${this.apiEnvironment.apiBase}/terms/accept`, {});
   }
+  */
 }
